refactor _get_span_coverage_data
and use it in test level coverage asserts
